requirements.txt
click
javalang
beautifulsoup4
scikit-learn
PyYAML
tqdm
networkx
platformdirs
